Address CR feedback.
