+++ assets/js/modules/analytics-4/components/audience-segmentation/dashboard/AudienceTilesWidget/index.stories.js
@@ -474,6 +474,16 @@ Loading.args = {
 			.startResolution( 'getReport', [ reportOptions ] );
 	},
 };
+Loading.decorators = [
+	( Story ) => {
+		// Ensure the animation is paused for VRT tests to correctly capture the loading state.
+		return (
+			<div className="googlesitekit-vrt-animation-paused">
+				<Story />
+			</div>
+		);
+	},
+];
 Loading.scenario = {};
 
 export default {
