+++ assets/js/modules/analytics-4/components/audience-segmentation/dashboard/AudienceTilesWidget/index.test.js
@@ -31,8 +31,137 @@ import {
 import { CORE_USER } from '../../../../../../googlesitekit/datastore/user/constants';
 import { withWidgetComponentProps } from '../../../../../../googlesitekit/widgets/util';
 import { availableAudiences } from '../../../../datastore/__fixtures__';
-import { MODULES_ANALYTICS_4 } from '../../../../datastore/constants';
-import { provideAudienceTilesMockReport } from '../../../../utils/data-mock';
+import {
+	MODULES_ANALYTICS_4,
+	DATE_RANGE_OFFSET,
+} from '../../../../datastore/constants';
+import { getAnalytics4MockResponse } from '../../../../utils/data-mock';
+
+/**
+ * Generates mock response for audience tiles component.
+ *
+ * @since n.e.x.t
+ *
+ * @param {Object}        registry            Data registry object.
+ * @param {Array<string>} configuredAudiences Array of audience resource names.
+ */
+function provideAudienceTilesMockReport( registry, configuredAudiences ) {
+	const dates = registry.select( CORE_USER ).getDateRangeDates( {
+		offsetDays: DATE_RANGE_OFFSET,
+		compare: true,
+	} );
+
+	const { startDate, endDate } = dates;
+
+	const reportOptions = {
+		dimensions: [ { name: 'audienceResourceName' } ],
+		metrics: [
+			{ name: 'totalUsers' },
+			{ name: 'sessionsPerUser' },
+			{ name: 'screenPageViewsPerSession' },
+			{ name: 'screenPageViews' },
+		],
+	};
+
+	const options = {
+		...dates,
+		...reportOptions,
+		dimensionFilters: {
+			audienceResourceName: configuredAudiences,
+		},
+	};
+
+	const reportData = getAnalytics4MockResponse( options );
+
+	registry.dispatch( MODULES_ANALYTICS_4 ).receiveGetReport( reportData, {
+		options,
+	} );
+
+	registry
+		.dispatch( MODULES_ANALYTICS_4 )
+		.finishResolution( 'getReport', [ options ] );
+
+	const totalPageviewsReportOptions = {
+		startDate,
+		endDate,
+		metrics: [ { name: 'screenPageViews' } ],
+	};
+
+	const totalPageviewsReportData = getAnalytics4MockResponse( options );
+
+	registry
+		.dispatch( MODULES_ANALYTICS_4 )
+		.receiveGetReport( totalPageviewsReportData, {
+			options: totalPageviewsReportOptions,
+		} );
+
+	registry
+		.dispatch( MODULES_ANALYTICS_4 )
+		.finishResolution( 'getReport', [ totalPageviewsReportOptions ] );
+
+	const topCitiesReportOptions = {
+		startDate,
+		endDate,
+		dimensions: [ 'city' ],
+		metrics: [ { name: 'totalUsers' } ],
+		orderby: [
+			{
+				metric: {
+					metricName: 'totalUsers',
+				},
+				desc: true,
+			},
+		],
+		limit: 3,
+	};
+
+	const topContentReportOptions = {
+		startDate,
+		endDate,
+		dimensions: [ 'pagePath' ],
+		metrics: [ { name: 'screenPageViews' } ],
+		orderby: [ { metric: { metricName: 'screenPageViews' }, desc: true } ],
+		limit: 3,
+	};
+
+	const topContentPageTitlesReportOptions = {
+		startDate,
+		endDate,
+		dimensions: [ 'pagePath', 'pageTitle' ],
+		metrics: [ { name: 'screenPageViews' } ],
+		orderby: [ { metric: { metricName: 'screenPageViews' }, desc: true } ],
+		limit: 15,
+	};
+
+	[
+		topCitiesReportOptions,
+		topContentReportOptions,
+		topContentPageTitlesReportOptions,
+	].forEach( ( value ) => {
+		configuredAudiences.forEach( ( audienceResourceName ) => {
+			const individualReportOptions = {
+				...value,
+				dimensionFilters: {
+					audienceResourceName,
+				},
+			};
+
+			const individualReportData = getAnalytics4MockResponse(
+				individualReportOptions
+			);
+
+			registry
+				.dispatch( MODULES_ANALYTICS_4 )
+				.receiveGetReport( individualReportData, {
+					options: individualReportOptions,
+				} );
+
+			registry
+				.dispatch( MODULES_ANALYTICS_4 )
+				.finishResolution( 'getReport', [ individualReportOptions ] );
+		} );
+	} );
+}
 
 describe( 'AudienceTilesWidget', () => {
 	let registry;
