+++ assets/js/modules/analytics-4/utils/data-mock.js
@@ -29,7 +29,7 @@ import { map, reduce, take, toArray, mergeMap } from 'rxjs/operators';
 /**
  * Internal dependencies
  */
-import { DATE_RANGE_OFFSET, MODULES_ANALYTICS_4 } from '../datastore/constants';
+import { MODULES_ANALYTICS_4 } from '../datastore/constants';
 import {
 	isValidDateString,
 	stringifyObject,
@@ -37,7 +37,6 @@ import {
 } from '../../../util';
 import { isValidDimensionFilters } from './report-validation';
 import { isValidPivotsObject } from './report-pivots-validation';
-import { CORE_USER } from '../../../googlesitekit/datastore/user/constants';
 
 export const STRATEGY_CARTESIAN = 'cartesian';
 export const STRATEGY_ZIP = 'zip';
@@ -1021,133 +1020,4 @@ export function provideAnalytics4MockPivotReport( registry, options ) {
 		.receiveGetReport( getAnalytics4MockPivotResponse( options ), {
 			options,
 		} );
-}
-
-/**
- * Generates mock response for audience tiles component.
- *
- * @since n.e.x.t
- *
- * @param {Object}        registry            Data registry object.
- * @param {Array<string>} configuredAudiences Array of audience resource names.
- */
-export function provideAudienceTilesMockReport(
-	registry,
-	configuredAudiences
-) {
-	const dates = registry.select( CORE_USER ).getDateRangeDates( {
-		offsetDays: DATE_RANGE_OFFSET,
-		compare: true,
-	} );
-
-	const { startDate, endDate } = dates;
-
-	const reportOptions = {
-		dimensions: [ { name: 'audienceResourceName' } ],
-		metrics: [
-			{ name: 'totalUsers' },
-			{ name: 'sessionsPerUser' },
-			{ name: 'screenPageViewsPerSession' },
-			{ name: 'screenPageViews' },
-		],
-	};
-
-	const options = {
-		...dates,
-		...reportOptions,
-		dimensionFilters: {
-			audienceResourceName: configuredAudiences,
-		},
-	};
-
-	const reportData = getAnalytics4MockResponse( options );
-
-	registry.dispatch( MODULES_ANALYTICS_4 ).receiveGetReport( reportData, {
-		options,
-	} );
-
-	registry
-		.dispatch( MODULES_ANALYTICS_4 )
-		.finishResolution( 'getReport', [ options ] );
-
-	const totalPageviewsReportOptions = {
-		startDate,
-		endDate,
-		metrics: [ { name: 'screenPageViews' } ],
-	};
-
-	const totalPageviewsReportData = getAnalytics4MockResponse( options );
-
-	registry
-		.dispatch( MODULES_ANALYTICS_4 )
-		.receiveGetReport( totalPageviewsReportData, {
-			options: totalPageviewsReportOptions,
-		} );
-
-	registry
-		.dispatch( MODULES_ANALYTICS_4 )
-		.finishResolution( 'getReport', [ totalPageviewsReportOptions ] );
-
-	const topCitiesReportOptions = {
-		startDate,
-		endDate,
-		dimensions: [ 'city' ],
-		metrics: [ { name: 'totalUsers' } ],
-		orderby: [
-			{
-				metric: {
-					metricName: 'totalUsers',
-				},
-				desc: true,
-			},
-		],
-		limit: 3,
-	};
-
-	const topContentReportOptions = {
-		startDate,
-		endDate,
-		dimensions: [ 'pagePath' ],
-		metrics: [ { name: 'screenPageViews' } ],
-		orderby: [ { metric: { metricName: 'screenPageViews' }, desc: true } ],
-		limit: 3,
-	};
-
-	const topContentPageTitlesReportOptions = {
-		startDate,
-		endDate,
-		dimensions: [ 'pagePath', 'pageTitle' ],
-		metrics: [ { name: 'screenPageViews' } ],
-		orderby: [ { metric: { metricName: 'screenPageViews' }, desc: true } ],
-		limit: 15,
-	};
-
-	[
-		topCitiesReportOptions,
-		topContentReportOptions,
-		topContentPageTitlesReportOptions,
-	].forEach( ( value ) => {
-		configuredAudiences.forEach( ( audienceResourceName ) => {
-			const individualReportOptions = {
-				...value,
-				dimensionFilters: {
-					audienceResourceName,
-				},
-			};
-
-			const individualReportData = getAnalytics4MockResponse(
-				individualReportOptions
-			);
-
-			registry
-				.dispatch( MODULES_ANALYTICS_4 )
-				.receiveGetReport( individualReportData, {
-					options: individualReportOptions,
-				} );
-
-			registry
-				.dispatch( MODULES_ANALYTICS_4 )
-				.finishResolution( 'getReport', [ individualReportOptions ] );
-		} );
-	} );
 }
